pyproject.toml
[project]
name = "mtg-card-aggregator"
version = "0.0.1"
requires-python = ">= 3.12"
dependencies = [
    "typer",
    "rich",
    "ijson",
    "requests",
    "jinja2",
    "beautifulsoup4>=4.13.3",
    "pyyaml>=6.0.2",
]

[tool.uv]
package = false
